+++ HISTORY.rst
@@ -12,6 +12,7 @@
 - Make document version OCR submit view messages translatable.
 - Make file caching purge view messages translatable.
 - Make document file metadata submit view messages translatable.
+- Fix asset transformations hash calculation.
 
 4.2.4 (2022-04-29)
 ==================